requirements.txt
bleach==1.5.0
certifi==2019.9.11
cffi==1.13.2
Click==7.0
cycler==0.10.0
decorator==4.4.1
enum34==1.1.6
Flask==1.0
gevent==1.3.0
greenlet==0.4.15
h5py==2.7.1
html5lib==0.9999999
imgaug==0.2.5
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.1.3
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.22.0
pathlib2==2.3.2
Pillow==5.0.0
protobuf==3.10.0
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.13.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
Werkzeug==0.15.4
wincertstore==0.2
